refactor accordion and accordion-item blocks
