+++ src/blocks/accordion/accordion-item/edit.js
@@ -28,7 +28,7 @@ const TEMPLATE = [ [ 'core/paragraph', { placeholder: 'Add content...' } ] ];
 /**
  * Block edit function
  */
-class Edit extends Component {
+class edit extends Component {
 	render() {
 		const {
 			attributes,
@@ -104,4 +104,4 @@ class Edit extends Component {
 
 export default compose( [
 	applyWithColors,
-] )( Edit );
+] )( edit );
