+++ src/blocks/accordion/accordion-item/index.js
@@ -1,70 +1,34 @@
 /**
- * External dependencies
- */
-import classnames from 'classnames';
-
-/**
  * Internal dependencies
  */
 import './styles/editor.scss';
 import './styles/style.scss';
-import Edit from './components/edit';
+import edit from './edit';
+import save from './save';
 import icons from './../../../utils/icons';
+import metadata from './block.json';
 
 /**
  * WordPress dependencies
  */
 const { __ } = wp.i18n;
-const { RichText, InnerBlocks, getColorClassName } = wp.blockEditor;
 
 /**
  * Block constants
  */
-const name = 'accordion-item';
+const { name } = metadata;
 
 const title = __( 'Accordion Item' );
 
 const icon = icons.accordionItem;
 
-const keywords = [
-	__( 'tabs' ),
-	__( 'faq' ),
-	__( 'coblocks' ),
-];
-
-const blockAttributes = {
-	title: {
-		type: 'string',
-		selector: '.wp-block-coblocks-accordion__title',
-	},
-	open: {
-		type: 'boolean',
-		default: false,
-	},
-	backgroundColor: {
-		type: 'string',
-	},
-	textColor: {
-		type: 'string',
-	},
-	borderColor: {
-		type: 'string',
-	},
-	customBackgroundColor: {
-		type: 'string',
-	},
-	customTextColor: {
-		type: 'string',
-	},
-};
-
 const settings = {
 
-	title: title,
+	title,
 
 	description: __( 'Add collapsable accordion items to accordions.' ),
 
-	keywords: keywords,
+	keywords: [	__( 'tabs' ), __( 'faq' ), __( 'coblocks' )	],
 
 	parent: [ 'coblocks/accordion' ],
 
@@ -74,59 +38,11 @@ const settings = {
 		inserter: false,
 	},
 
-	attributes: blockAttributes,
+	attributes: metadata.attributes,
 
-	edit: Edit,
+	edit,
 
-	save( { attributes } ) {
-		const {
-			backgroundColor,
-			customBackgroundColor,
-			customTextColor,
-			open,
-			textColor,
-			borderColor,
-			title,
-		} = attributes;
-
-		const backgroundColorClass = getColorClassName( 'background-color', backgroundColor );
-		const textColorClass = getColorClassName( 'color', textColor );
-
-		const titleClasses = classnames(
-			'wp-block-coblocks-accordion-item__title', {
-				'has-background': backgroundColor || customBackgroundColor,
-				[ backgroundColorClass ]: backgroundColorClass,
-				'has-text-color': textColor || customTextColor,
-				[ textColorClass ]: textColorClass,
-			} );
-
-		const titleStyles = {
-			backgroundColor: backgroundColorClass ? undefined : customBackgroundColor,
-			color: textColorClass ? undefined : customTextColor,
-		};
-
-		const borderStyle = {
-			borderColor: borderColor ? borderColor : customBackgroundColor,
-		};
-
-		return (
-			<div>
-				{ ! RichText.isEmpty( title ) &&
-					<details open={ open }>
-						<RichText.Content
-							tagName="summary"
-							className={ titleClasses }
-							value={ title }
-							style={ titleStyles }
-						/>
-						<div className="wp-block-coblocks-accordion-item__content" style={ borderStyle }>
-							<InnerBlocks.Content />
-						</div>
-					</details>
-				}
-			</div>
-		);
-	},
+	save,
 };
 
 export { name, title, icon, settings };
