+++ src/blocks/accordion/block.json
@@ -1,0 +1,14 @@
+{
+	"name": "accordion",
+	"category": "coblocks",
+	"attributes": {
+		"count": {
+			"type": "number",
+			"default": "1"
+		},
+		"polyfill": {
+			"type": "boolean",
+			"default": false
+		}
+	}
+}
