+++ src/blocks/accordion/index.js
@@ -4,82 +4,44 @@
  */
 import './styles/editor.scss';
 import './styles/style.scss';
-import Edit from './components/edit';
+import edit from './edit';
 import icons from './../../utils/icons';
+import transforms from './transforms';
+import metadata from './block.json';
 
 /**
  * WordPress dependencies
  */
 const { __ } = wp.i18n;
-const { createBlock } = wp.blocks;
 const { InnerBlocks } = wp.blockEditor;
 
 /**
  * Block constants
  */
-const name = 'accordion';
+const { name } = metadata;
 
 const title = __( 'Accordion' );
 
 const icon = icons.accordion;
 
-const keywords = [
-	__( 'tabs' ),
-	__( 'faq' ),
-	__( 'coblocks' ),
-];
-
-const blockAttributes = {
-	count: {
-		type: 'number',
-		default: 1,
-	},
-	polyfill: {
-		type: 'boolean',
-		default: false,
-	},
-};
-
 const settings = {
 
-	title: title,
+	title,
 
 	description: __( 'Organize content within collapsable accordion items.' ),
 
-	keywords: keywords,
+	keywords: [	__( 'tabs' ), __( 'faq' ), __( 'coblocks' ) ],
 
-	attributes: blockAttributes,
+	attributes: metadata.attributes,
 
 	supports: {
 		align: [ 'wide', 'full' ],
 		html: false,
 	},
 
-	transforms: {
-		from: [
-			{
-				type: 'prefix',
-				prefix: ':accordion',
-				transform: function( content ) {
-					return createBlock( `coblocks/${ name }`, {
-						content,
-					} );
-				},
-			},
-			...[ 2, 3, 4, 5 ].map( ( count ) => ( {
-				type: 'prefix',
-				prefix: Array( count + 1 ).join( ':' ) + 'accordion',
-				transform( content ) {
-					return createBlock( `coblocks/${ name }`, {
-						content,
-						count,
-					} );
-				},
-			} ) ),
-		],
-	},
+	transforms,
 
-	edit: Edit,
+	edit,
 
 	save() {
 		return (
